requirement.txt
unstructured
tiktoken
pinecone-client
pypdf
langchain
langchain-community
transformers
pandas
numpy
python-dotenv